Rename is-object to isObject.
