+++ lib/lru/promote.js
@@ -4,7 +4,7 @@ var __tail = require("./../internal/tail");
 var __next = require("./../internal/next");
 var __prev = require("./../internal/prev");
 
-var isObject = require("./../support/is-object");
+var isObject = require("./../support/isObject");
 
 module.exports = function lruPromote(root, node) {
