+++ lib/set/sync.js
@@ -1,9 +1,9 @@
 var $error = require("./../types/error");
 var pathSyntax = require("falcor-path-syntax");
 var getType = require("./../support/get-type");
-var isObject = require("./../support/is-object");
 var isPathValue = require("./../support/is-path-value");
 var setJsonValuesAsJsonDense = require("./../set/set-json-values-as-json-dense");
+var isObject = require("./../support/isObject");
 
 module.exports = function setValueSync(pathArg, valueArg, errorSelectorArg, comparatorArg) {
