+++ lib/support/hasOwn.js
@@ -1,4 +1,4 @@
-var isObject = require("./is-object");
+var isObject = require("./isObject");
 var hasOwn = Object.prototype.hasOwnProperty;
 
 module.exports = function(obj, prop) {
